Added testingFx, Transaction, and Account
These are the class files from the original program I made. The interface is basic and Clunky but it can perform the necessary Operations. This is the effective back end of the project, while animationTesting is the front End. Transaction and Account are Aggregated into testingFx, where the main interface is. Use of data structures is simple and could use work in final product.